Update II - Hotel BT Website
Styled the navigation bar with coloring for text and background. Added the over flow rule so the border wouldn't override the text. The navigation menu is floated to the right while the logo is floated to the left. Added a hover state and  current class so the nav menu shows which page you are on. Added a container class as a way to move items when in the header border depending one the screen size.