Update III - Hotel BT
Added the Showcase section. It shows a faded out hotel room picture with basic hotel info and a button to the About page of the site.
Created 3 more utility classes for text-primary, lead, and button class. Text-primary is for assigning a main color to a word or phrase. Lead is for increasing font size for the lead wording for sections. Btn is for styling for buttons and a hover state for the button.

Styled the showcase section. Added the backgrounf image and give it some height. Added padding for spacing, aligned the text center and colored text. For heading and section text, increased font size, line height, bottom padding.