Update IV - Hotel BT
Created the Home Information section. Added a div for the image of the front of the hotel. Added another div fo the brief history of the hotel with title and a button to the About page.

Added the home-info id for setting the general height for the section. Created bg-dark for setting a drak background. Also added bg-light for a light background and bg-primary to set a background to the primary tan color. Added a hotel image for the info section, by floating it to the left and setting width and height requirements. Added the History of the Hotel section styles rules. Setting colors spacing center alignment. Set overflow to hidden so the user doesn;t have to scroll the section. Add btn-light class to have a light colored button.